move pre transform stuff to its own system
