+++ packages/engine/src/avatar/AvatarClientModule.ts
@@ -1,6 +1,7 @@
 import { Engine } from '../ecs/classes/Engine'
 import { initSystems } from '../ecs/functions/SystemFunctions'
 import { SystemUpdateType } from '../ecs/functions/SystemUpdateType'
+import ReferenceSpaceTransformSystem from '../transform/systems/ReferenceSpaceTransformSystem'
 import AnimationSystem from './AnimationSystem'
 import AvatarAnimationSystem from './AvatarAnimationSystem'
 import AvatarControllerSystem from './AvatarControllerSystem'
@@ -36,6 +37,12 @@ export function AvatarClientModule() {
       uuid: 'xre.engine.AvatarMovementSystem',
       type: SystemUpdateType.FIXED_EARLY,
       systemLoader: () => Promise.resolve({ default: AvatarMovementSystem })
+    },
+    {
+      uuid: 'xre.engine.ReferenceSpaceTransformSystem',
+      type: SystemUpdateType.UPDATE_LATE,
+      before: 'xre.engine.TransformSystem',
+      systemLoader: () => Promise.resolve({ default: ReferenceSpaceTransformSystem })
     },
     {
       uuid: 'xre.engine.AvatarAnimationSystem',
